+++ apps/web/app/page.tsx
@@ -1,5 +1,5 @@
 import Input from "@ui/src/components/common/Input";
-import DropdownTest from "@ui/src/components/common/Dropdown/DropdownTest";
+import DropdownExample from "@ui/src/components/common/Dropdown/Example/DropdownExample";
 
 export default function Home(): JSX.Element {
   return (
@@ -7,7 +7,7 @@ export default function Home(): JSX.Element {
       <div className="m-30 flex flex-col gap-20">
         <Input id="name" placeholder="참여자" />
         <Input id="name" placeholder="참여자" isError errorMessage="에러 예시" />
-        <DropdownTest />
+        <DropdownExample />
       </div>
     </div>
   );
